// src/app/common/models/HotelInfo.ts
/**
* HotelInfo info
*/
export class HotelInfo  {
 /** HotelInfo fields. */
 hotelId: number = 0;
 hotelStarRating: number = 0.0;

 hotelName: string = "";
 hotelDestination: string = "";
 hotelCity: string = "";
 hotelImageUrl: string = "";
 hotelStreetAddress: string = "";

 hotelInfositeUrl: string = "";
 hotelSearchResultUrl: string = "";

}